+++ crates/ty_ide/src/lib.rs
@@ -31,10 +31,9 @@ mod type_hierarchy;
 mod workspace_symbols;
 
 pub use all_symbols::{AllSymbolInfo, all_symbols};
-pub use call_hierarchy::{
-    CallHierarchyIncomingCall, CallHierarchyItem, CallHierarchyItemKind, CallHierarchyOutgoingCall,
-    call_hierarchy_incoming_calls, call_hierarchy_outgoing_calls, prepare_call_hierarchy,
-};
+pub use call_hierarchy::incoming_calls::{IncomingCall, incoming_calls};
+pub use call_hierarchy::outgoing_calls::{OutgoingCall, outgoing_calls};
+pub use call_hierarchy::{CallHierarchyItem, CallHierarchyItemKind, prepare_call_hierarchy};
 pub use code_action::{QuickFix, code_actions};
 pub use completion::{Completion, CompletionKind, CompletionSettings, completion};
 pub use doc_highlights::document_highlights;
